Core logic for separating the output markdown generation logic.
